Move fetcher path as ./lib
