+++ src/lib/api/contentful.ts
@@ -1,4 +1,4 @@
-import { fetcher } from "./fetcher";
+import { fetcher } from "@/src/lib/fetcher";
 import { Banner } from "@/src/type/contentful";
 import { SSG } from "@/src/constants/rendering-type";
